requirements.txt
boto3
pyspark
requests
python-dotenv
pytest